+++ RELEASE_NOTES.md
@@ -78,3 +78,11 @@
 
 #### 0.0.17 - 
 * Improvements to symbol accuracy w.r.t. type abbreviations 
+
+#### 0.0.18 - 
+* Add GetAllUsesOfAllSymbols and GetAllUsesOfAllSymbolsInFile
+
+#### 0.0.19 - 
+* Change return type of GetAllUsesOfSymbol, GetAllUsesOfAllSymbols and GetAllUsesOfAllSymbolsInFile to FSharpSymbolUse
+* Add symbol uses when an abstract member is implemented.
+